nmake: allow building with a custom C compiler (#181)
* xpath: fix non-matching function name in xmlSecError() call

* nmake: allow specifying a custom compiler

clang-cl.exe is a drop-in replacement for cl.exe and has some useful additional
warnings which cl.exe lacks.

* mscng: fix warnings found by clang-cl

-Werror,-Wunused-value is great.

+++ src/mscng/certkeys.c
@@ -908,7 +908,7 @@ xmlSecMSCngKeyDataDsaDebugDump(xmlSecKeyDataPtr data, FILE* output) {
     xmlSecAssert(output != NULL);
 
     fprintf(output, "=== rsa key: size = %d\n",
-            xmlSecMSCngKeyDataDsaGetSize(data));
+            (int)xmlSecMSCngKeyDataDsaGetSize(data));
 }
 
 static void xmlSecMSCngKeyDataDsaDebugXmlDump(xmlSecKeyDataPtr data, FILE* output) {
@@ -916,7 +916,7 @@ static void xmlSecMSCngKeyDataDsaDebugXmlDump(xmlSecKeyDataPtr data, FILE* outpu
     xmlSecAssert(output != NULL);
 
     fprintf(output, "<DSAKeyValue size=\"%d\" />\n",
-            xmlSecMSCngKeyDataDsaGetSize(data));
+            (int)xmlSecMSCngKeyDataDsaGetSize(data));
 }
 
 static int
@@ -1120,7 +1120,7 @@ xmlSecMSCngKeyDataRsaDebugDump(xmlSecKeyDataPtr data, FILE* output) {
     xmlSecAssert(output != NULL);
 
     fprintf(output, "=== rsa key: size = %d\n",
-            xmlSecMSCngKeyDataRsaGetSize(data));
+            (int)xmlSecMSCngKeyDataRsaGetSize(data));
 }
 
 static void xmlSecMSCngKeyDataRsaDebugXmlDump(xmlSecKeyDataPtr data, FILE* output) {
@@ -1128,7 +1128,7 @@ static void xmlSecMSCngKeyDataRsaDebugXmlDump(xmlSecKeyDataPtr data, FILE* outpu
     xmlSecAssert(output != NULL);
 
     fprintf(output, "<RSAKeyValue size=\"%d\" />\n",
-            xmlSecMSCngKeyDataRsaGetSize(data));
+            (int)xmlSecMSCngKeyDataRsaGetSize(data));
 }
 
 static int
@@ -1587,7 +1587,7 @@ xmlSecMSCngKeyDataEcdsaDebugDump(xmlSecKeyDataPtr data, FILE* output) {
     xmlSecAssert(output != NULL);
 
     fprintf(output, "=== rsa key: size = %d\n",
-            xmlSecMSCngKeyDataEcdsaGetSize(data));
+            (int)xmlSecMSCngKeyDataEcdsaGetSize(data));
 }
 
 static void xmlSecMSCngKeyDataEcdsaDebugXmlDump(xmlSecKeyDataPtr data, FILE* output) {
@@ -1595,7 +1595,7 @@ static void xmlSecMSCngKeyDataEcdsaDebugXmlDump(xmlSecKeyDataPtr data, FILE* out
     xmlSecAssert(output != NULL);
 
     fprintf(output, "<ECDSAKeyValue size=\"%d\" />\n",
-            xmlSecMSCngKeyDataEcdsaGetSize(data));
+            (int)xmlSecMSCngKeyDataEcdsaGetSize(data));
 }
 
 static xmlSecKeyDataKlass xmlSecMSCngKeyDataEcdsaKlass = {
